Improved imput handling and separated nmap function(s) into separate file!
