+++ Cargo.toml
@@ -8,3 +8,4 @@ rand = "0.8"
 tokio = { version = "1", features = ["full"] }
 colored = "2"
 nix = "0.26"
+anyhow = "1.0"
