+++ src/main.rs
@@ -1,14 +1,46 @@
 mod file_utils;
+mod nmap;
+
 use colored::*;
-use file_utils::{check_directory_exists, create_file};
+use file_utils::check_directory_exists;
+use nmap::map_target;
 use rand::Rng;
 use std::io::{self, Write};
-use std::process::Command;
+use std::net::IpAddr;
+use std::str::FromStr;
 use tokio::time::{sleep, Duration};
+
+fn get_input(prompt: &str) -> String {
+    let mut input = String::new();
+    print!("{}", prompt);
+    io::stdout().flush().unwrap();
+    io::stdin()
+        .read_line(&mut input)
+        .expect(&"Failed to read input.".red());
+    input.trim().to_string()
+}
+
+fn validate_ip(ip: &str) -> bool {
+    match IpAddr::from_str(ip) {
+        Ok(_) => true,
+        Err(_) => false,
+    }
+}
+
+fn validate_port(port: &str) -> bool {
+    port.parse::<u16>().is_ok()
+}
 
 fn generate_number(min: i32, max: i32) -> i32 {
     let mut rng = rand::thread_rng();
     rng.gen_range(min..=max)
+}
+
+fn format_seconds(seconds: u64) -> String {
+    let hours = seconds / 3600;
+    let minutes = (seconds % 3600) / 60;
+    let seconds = seconds % 60;
+    format!("{:02}:{:02}:{:02}", hours, minutes, seconds)
 }
 
 async fn wait(seconds: u64) {
@@ -24,73 +56,6 @@ async fn wait(seconds: u64) {
     println!("{}", "\rRunning Next Scan...".yellow());
 }
 
-fn format_seconds(seconds: u64) -> String {
-    let hours = seconds / 3600;
-    let minutes = (seconds % 3600) / 60;
-    let seconds = seconds % 60;
-    format!("{:02}:{:02}:{:02}", hours, minutes, seconds)
-}
-
-fn map_target(ip: &str, port: &str, scan_type: Option<&str>, output_prefix: &str) {
-    let scan_type_str = match scan_type {
-        Some("-sS") => "SYN",
-        Some("-sF") => "FIN",
-        Some("-sN") => "NULL",
-        Some("-sT") => "CONNECT",
-        None => "REGULAR",
-        Some(_) => "",
-    };
-
-    let file_name = format!(
-        "{PREFIX}_{TYPE}",
-        PREFIX = output_prefix,
-        TYPE = scan_type_str
-    );
-    match create_file(&file_name) {
-        Ok(mut file) => {
-            let mut command = Command::new("sudo");
-
-            command.arg("nmap");
-            command.arg(ip);
-            if let Some(f) = scan_type {
-                command.arg(f);
-            }
-            if !port.trim().is_empty() {
-                command.arg("-p").arg(port);
-            }
-
-            let output = command.output().expect(&"Failed to execute command".red());
-
-            if output.status.success() {
-                file.write_all(&output.stdout)
-                    .expect(&"Failed to write output to file".red());
-                println!(
-                    "\nnmap {} {} {}",
-                    ip,
-                    scan_type.unwrap_or(""),
-                    if port.trim().is_empty() {
-                        "".to_string()
-                    } else {
-                        format!("-p {}", port)
-                    }
-                );
-
-                println!("{}", format!("Output Stored In: {}.txt", file_name).green());
-            } else {
-                eprintln!(
-                    "{}",
-                    format!(
-                        "Nmap command failed with error: {}",
-                        String::from_utf8_lossy(&output.stderr)
-                    )
-                    .red()
-                );
-            }
-        }
-        Err(e) => eprintln!("{}", format!("Error creating file: {}", e).red()),
-    }
-}
-
 fn main() {
     match check_directory_exists("map_data") {
         Ok(map_data_dir) => println!(
@@ -100,44 +65,24 @@ fn main() {
         Err(e) => eprintln!("Error: {}", e),
     }
 
-    let mut target_ip = String::new();
-    print!("\nTarget IP: ");
-    io::stdout().flush().unwrap();
-    io::stdin()
-        .read_line(&mut target_ip)
-        .expect(&"Failed to read input.".red());
+    let target_ip = get_input("\nTarget IP: ");
+    if !validate_ip(&target_ip) {
+        eprintln!("Invalid IP Address...");
+        return;
+    }
+    let target_port = get_input("Target Port (ENTER for All): ");
+    if !target_port.is_empty() && !validate_port(&target_port) {
+        eprintln!("Invalid Port...");
+        return;
+    }
+    let file_prefix = get_input("Output File Prefix: ");
+    let mode_choice = get_input("Choose Mode (Stealth / Fast): ").to_lowercase();
 
-    let mut target_port = String::new();
-    print!("Target Port (Enter for ALL): ");
-    io::stdout().flush().unwrap();
-    io::stdin()
-        .read_line(&mut target_port)
-        .expect(&"Failed to read input.".red());
-
-    let mut file_prefix = String::new();
-    print!("Output File Prefix: ");
-    io::stdout().flush().unwrap();
-    io::stdin()
-        .read_line(&mut file_prefix)
-        .expect(&"Failed to read input.".red());
-
-    let mut mode_choice = String::new();
-    print!("Choose Mode (Stealth / Fast): ");
-    io::stdout().flush().unwrap();
-    io::stdin()
-        .read_line(&mut mode_choice)
-        .expect(&"Failed to read input.".red());
-
-    let mode_choice = mode_choice.trim().to_lowercase();
     let (min_wait_time, max_wait_time) = if mode_choice == "fast" {
         (0, 300)
     } else {
         (1800, 7200)
     };
-
-    let target_ip = target_ip.trim();
-    let target_port = target_port.trim();
-    let file_prefix = file_prefix.trim();
 
     let rt = tokio::runtime::Runtime::new().unwrap();
     let scan_flags = [None, Some("-sS"), Some("-sF"), Some("-sN"), Some("-sT")];
@@ -145,7 +90,7 @@ fn main() {
     for i in 0..=4 {
         let flag = scan_flags.get(i % scan_flags.len()).copied().flatten();
 
-        map_target(target_ip, target_port, flag, file_prefix);
+        let _ = map_target(&target_ip, &target_port, flag, &file_prefix);
 
         let wait_time = generate_number(min_wait_time, max_wait_time);
         rt.block_on(wait(wait_time.try_into().unwrap()));
